cleaned code, added 'support absent' columns
